move sanity data into @app-boilerplate
