+++ packages/@app-boilerplate/sanity-data/.babelrc
