+++ packages/@app-boilerplate/sanity-data/README.md
@@ -1,4 +1,4 @@
-# @utils/sanity-data
+# @app-boilerplate/sanity-data
 
 Contains data that needs to be prefetched by Sanity. Typically it's for things that need to be fetched outside of Gatsby, otherwise Gatsby's Sanity plugin should handle it.
