+++ packages/@app-boilerplate/sanity-data/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@utils/sanity-data",
+  "name": "@app-boilerplate/sanity-data",
   "version": "0.0.0",
   "private": true,
   "main": "index.js",
